many fixes and improvements to setup routine: - fixed bug that password was saved multiple times in db - fixed bug that a blank password were possible - implemented db_updater in setup-routine. setup detects changes made in dbupdate.php and provides an option to update your database - all language values displayed should now come from the lang_setup-files
when you have to change to database please use the script dbupdate.php in future.
Just increment the last version number you find in this format: <#3>
Behind the number put your SQL-Code or PHP-Code encapsulated with <?php ?>
Rules & Limitations:
You may only use the short syntax in SQL. Example: instead of
INSERT INTO table ('column1') VALUES ('val1');
use
INSERT INTO table VALUES ('val1');

Don't forget the semicolon.

When using PHP-Code your db-handler is '$this->db' ! and not '$ilias->db' as in the previous versions
of ILIAS 3. The db object is an instance of PEAR::DB.
Don't use the following varnames as their are already used in the DBUdate class:
$sql, $php, $update, $mode, $res, $row, $i, $nr


git-svn-id: http://svn.ilias.de/svn/ilias/trunk@570 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5