Corrected character data handling in function handlerCharacterData() so that <title> of <item> and <organization> are set properly.
The handlerCharacterData()  function is called by the xml parser for every
token of the content, so it is called for every piece of "string" and for
every "xml entity" in the content, so these have to be concatenated.

Note:
Here we do a database update for every call of handlerCharacterData().
This may become a performance issue if the content contains many
xml entities. Then the concatenation should be done in a private
variable and the db update done once in the handlerEndTag()  function.


git-svn-id: http://svn.ilias.de/svn/ilias/trunk@4462 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5