Restricted sendMessage to non participating chat
It was possible to send the window.il.Chat.sendMessage command from console and send messages to non participating chats by brute forcing the chat id. After this change only participating users are allowed to send messages to a chat. Other users, knowing the conversation id, but not participating are disallowed to send messages.