Fixed missing red badges counter after messages on offline in groupchat
There was an error, where users which have been added to a groupconversation while offline did not get notified about new messages after login. Problem was the missing activity entry which causes the query to return 0 until opening the chat manually once. It has been changed to track an faked activity on invite for each user in the chat.