Improved Proxy-URLS for Chatserver
In the Chatserver configuration there is an option to set a specific URL for the connection between ILIAS -> Server and the Client -> Server. This is used if the server is configured behind an reverse proxy and if the URLs how the server can be reached to the servers listening ip-address/FQN. Until now, the proxy paths uses the protocol defined in the configuration. But with this, it was impossible to create an HTTPS reverse proxy but running the chatserver without HTTP. This is now more flexible.