ilCtrl: removed forwards and related variables.
The ilCtrl::forwards method is private, returns no values and is only called in
one location. The member variables parent and forward are only used in the
forwards method. The two members are not used anywhere in the codebase as
`grep -r "lCtrl->forward" .` and `grep -r "ilCtrl->parent" .` reveal. So we can
just dump this.