5_2_refactor_svy_compulsory_bugfix_0019478
In my opinion we have all we need in the svy_question table, where we
store: question id, the origin of this question and if it's compulsory
or not. We don't need one table only to store if this question is
compulsory or not. The motivation of this is also fix this bug: 0019478