Use the newer version of jquery.nestable.js
In the "newer" version, this jquery plugin supports mouse and touch events.
The "older" version decided at the begin, if the used device has touch or not and set the events for this case. If, for example, the laptop has a touchscreen, it was not possible to drag and drop elements with the mouse.