Set Table ID before calling parent constructor.
I didn't mark this as a bug in Mantis but I did notice this and am proposing a fix. 

If you call parent::__construct before setting this table's ID determineLimit is called before setPrefix in ilTable2GUI so determineLimit never actually finds the GET var and row limiting in this table never ends up working.