Bugfix 0022151 - Fixed css caching in custom skins
To load the css from file instead of cache, just add a file
called "version" to your custom skins root folder and insert
a number (i.e. a skin version).
If this file does not exists, ilias behaves like before. If
the file exists but is empty, also nothing new happens. If
the file exists and has content, the content will be added to
the "?ver=" string,
Spaces and dots in the "skin version" will be replaced with
hyphens, like it is done with the standard ilias version.
example (replacing): "0.1.2" => "0-1-2"
example (?ver):
?ver=5-0-23-2017-07-19 => ?ver=5-0-23-2017-07-19-0-1-2