Bugfix #0022151 - Fixed css caching in custom skins
With this bugfix, ilias uses the version of template.xml
to append the "?ver" parameter on the styles css src-string.
The versions dots will be replaced by hyphens, before
appending the "?ver" paramter.
In case of delos skin is in use, the string will not be
modified. In case you want not to use this feature just add
"$Id$" as version, like it is done at the delos template.xml.
Skins, that have the version "1", like before this bugfix
suggested, the number will simply count up. You may edit
this attribute any time you need in your skins template.xml.
Skins, created by the UI skin editor, will automatically get
the version "0.1" on creation.
The automatic update always count up the last part of the
version (0.[1]). The version may be any number of parts,
separated by dots. Example: 0.1.0.0.[2]
The skin version is shown, but not editable, at the skin
settings inside the UI skin editor. Any successfully save
in the skins' main, less, color or icon settings, will
trigger the version update. So we are able to guarantee the
upadete on any change of the skin.

Bugfix of https://ilias.de/mantis/view.php?id=22151