Bugfix #0022151 - Fixed css caching in custom skins
With this bugfix, ilias uses the version of template.xml
to append the "?ver" parameter on the styles css src-string.
The versions dots will be replaced by hyphens, before
appending the "?ver" paramter.
In case of delos skin is in use, the string will not be
modified. In case you want not to use this feature just add
"$Id$" as version, like it is done at the delos template.xml.
You may edit this attribute any time you need in your skins
template.xml.

Bugfix of https://ilias.de/mantis/view.php?id=22151