Moved the DiskQuota classes in WebDAV-folder to DiskQuota-folder
There is already a DiskQuota-folder but with only 1 class. All other DiskQuota classes are in WebDAV. So I moved the DiskQuota classes form the WebDAV-folder to the DiskQuota-folder.

Additional, I changed all the includes to point at the new DiskQuota-folder instead of the WebDAV-folder. I added a Service.xml file to the DiskQuota. And I moved the DiskQuota templates from the WebDAV-folder to the DiskQuota-folder and changed the template includes.