UI: introduced ilGlobalTemplate
With this commit ilGlobalTemplate is introduced as a plain copy of ilTemplate.
Also usages of ilTemplate that are considered a usage of the "global functionality"
of ilTemplate are replaced with ilGlobalTemplate. A usage of ilTemplate was
considered a usage of the "global functionality" if either a template was loaded
that contains a `<html>`-tag or the `show`-method of the template was called.
Adjusted ilTemplate-typehints to ilGlobalTemplate-typehints on a best effort
base. Updated classmap in autoloader.