Language: fixed DB-usage in ilLanguage::__destruct
The usage of the database-dependency in the __destruct-method implicitely
introduces this dependency to all users of ilLanguage, even if they do not
use any method of ilLanguage in some (test-)scenario.

To fix this, the a (deprecated) check for said dependency was introduced.
This will lead to a new dicto-violation. The underlying problem should thus
be fixed by a proper solution, e.g. removing the language-log-functionality
or moving it to a decorator for the ilLanguage class which only is wrapped
around ilLanguage in the initialisation when language logging is activated.