UI: ilIndependentTemplate uses ilGlobalTemplate
ilIndependentTemplate is used in the automated tests of the UI-framework
and codes around the fact, that ilTemplate depends on an initialized ILIAS.
This got even messier when ilGlobalTemplate was introduced. There seems to
be some light at the end of the tunnel, since it should be possible to
get rid of more ILIAS-dependencies of ilTemplate soon.