[Fix 23357] Notification when survey was finished was always English
The method getUserIdsByEmail was returning  user logins instead of user ids. I changed it to return ids.
I created new method getUserLoginsByEmail to get the array of logins.
The non static method getUserIdByEmail was used only in one test and it was returning the first login found. I removed this method and use getUserIdByLogin instead.