Mail/NotificationCenter: Handle close event for mail notifications
This commit handles the 'Close' event for mail notification
center items. To let the user 'mute' the mail notification items,
the current timestamp is stored as a user reference whenever the
'Close' event is raised. The timestamp is used as a left interval
for subsequent renderings of the mail notification items.