BuddySystem/NotificationCenter: Handle close event for new requests
This commit handles the 'Close' event for contact request notification
center items. To let the user 'mute' the these notification items,
the current timestamp is stored as a user reference whenever the
'Close' event is raised. The timestamp is used as a left interval
for subsequent renderings of the notification items.