UI, Mainbar:  Revision
The goal is to have the entries of Mainbar, including subslates, adressable;
This opens the possibility to store engagement-states in a cookie.

The JS is way more structured and split up in model/renderer/persistence,
and the rendering is weeded out.
Slates are enumerated, and their signals are more aligned (note the public addOption-function in Signal).