UI/Select Input Field: fix #27837
The trigger here was that there are two forms on the same side, one with an
required input, the other without. A change in the first select triggered
an error in the second. Still, this should not happen, exceptions should
not be thrown at users.

The root cause for this is, that the empty string that is received from the
client is indeed a valid client side value, even if the field is required.
The "required" condition is evaluated later on.