UI/Numeric Input Field: fix #27555
Before this commit we used the default error message for the constraint in the
Numeric Input Field, which isn't really helpful to users, as pointed out in #27666.
The default message is replaced by a custom message. The message is created early
in the processing and the language module 'ui' isn't already loaded then, if the
module is loaded in the renderer. Thus, the loading of the language module is moved
to the initialisation of the UI framework. We assume, that we need the module when
someone requires the UI factory.