Service/Component: purge plugin ctrl-structure before reloading it
This is suppposed to fix https://mantis.ilias.de/view.php?id=28284 by making
the control structure loading for plugins more robust. Without the change,
updating the plugin just writes ctrl-structure into the database without
purging existing ctrl-structure entries for the plugin. This leads to the
problem described in the bug, when the structure of an non-installed plugin
is loaded. Since a ctrl-structure load for the plugin should load all entries
for that plugin anyway, we could purge existing entries before.