UICore: improve handling of paths in reader
`ilCtrlStructure::normalizePath` did not take into account that the same directory
might be referenced by different names, e.g., as in this case a relative vs. an
absolute name or via a symlink.

This improves this behaviour by adding `realpath` as a normalization measure for the
path. The explicit setting of `"."` as path to be scanned in the setup objective is
superfluous too.

This fixes two problems:

https://mantis.ilias.de/view.php?id=28284
https://mantis.ilias.de/view.php?id=28777