LSO: make polling intervall configurable
This makes the polling interval for the learning progress of legacy objects in the
LSO configurable for administrators and hence helps by improving the behaviour
documented in https://mantis.ilias.de/view.php?id=28840. This introduces a new
GUI in the administration that exposes the interval. The changes have been
discussed in PR 2876.