Setup: removed Agent::getConfigInput
This method initially was meant to be used in a WebGUI for the setup. This
won't be implemented for a foreseeable future and, if it will be implemented,
won't need an elaborate mechanism to input configuration for the configs.
The method was never really implemented or used, so can be removed without
any further ado.