Update class.ilObjMailGUI.php
Added autocomplete-disabling on mail_smtp_user and mail_smtp_password to avoid a pitfall.
If an user has saved his credientials for Ilias and those fields are supposed to be empty, the browser autocompletion will fill in the saved user-credentials and will set this data as mail-credentials. This will result in smtp connection errors.