Fix test result update after scoring system change (Mantis #29735)
When the scoring system is changed from “Incomplete or Partly Wrong
Answers Score Points” to “Only Correct and Complete Answers Score
Points”, the test results have to be recalculated.

Ensure this recalculation actually happens.

After the user confirms that recalculation is wanted, the form data is
saved. This invalidates one of the conditions of the recalculation
requirement. Of course, recalculation shall still be executed; this is
why we evaluate whether to recalculate _before_ saving the form
settings.

Fixes <https://mantis.ilias.de/view.php?id=29735>.