Setup FileSystem: 30122 precondition respects update and install
After the first PR we now noticed that a fresh installation would run into
an error 'Path is empty'.
This depends on a missing environment resource at this point, that is
used in a precondition of ilFileSystemComponentDataDirectoryCreatedObjective.
So we have to distinguish between a fresh ILIAS installation and an
already existing ILIAS installation at this precondition.

In case of a fresh installation we have to ensure all file system
directories are created. We use the config for file system from the
environment.

In case of an already installed ILIAS we can use the ini files.