Make reset notification work from table
Administrative notifications can be reset so that they are shown to all
users again, even when the users have already dismissed them. The reset
function in the list of all administrative notifications erroneously
called the delete function due to a copy and paste mistake.

Fix the action button entry to use the proper modal.