Fix order of questions not always being random
The order of questions pulled from a pool should always be random. This is not
the case when all questions from a pool are to be used, as @nicoroeser reported
to me. In this case all questions are pulled but they are sorted and not random
because `array_rand` returns a sorted array of keys and the questions are not
shuffled afterwards then. Thus, shuffling once at the end of `performBuild` is
needed. This also means that questions are shuffled twice in the first case
(questions still need to be pulled by `fetchQuestionsFromStageRandomly`), but
this should not pose a serious concern.