Improve wording for answering limitation (Mantis #31011)
This improves the instruction displayed when a multiple choice question
(with multiple selections) has a limitation configured on how many
solutions can be chosen.

Many language strings made users think that _exactly_ the given number
of solutions had to be ticked. But the setting is about the _maximum_
number, so the message should be as well.

This commit handles the language strings in English and German. The
German translation had already been better; this commit further improves
it. Other translations still need to be adjusted accordingly.

Fixes Mantis #31011 <https://mantis.ilias.de/view.php?id=31011>.