Survey: remove remaining code related to PhantomJS
PhantomJS [0] development has been suspended since March 2018. [1]
ILIAS has fully removed PhantomJS as a dependency in release 7. [2]

Mantis issue 28709 [3] was created because PDF export did not work in
the Survey module. (Historical note: this was likely related to commit
5afa87e4982ef1c9a6bcf339904a07b7acac8ad3, which incompletely merged the
Survey contributions from BGHW which enable PhantomJS support. The merge
lacks the updates to libs/composer/composer.json which install
PhantomJS.)

Mantis issue 28709 is fixed by commits
8b58ba9b242f8525bf668fda889e97f8b4fba338 (for release 6),
901c9bea182d3292036c2e796915147e364e8b61 (for release 7), and
433d5fec6edae55476311c96a5cd0456bda6a99a (for release 8 and later).

These commits do not fully remove the PhantomJS-related code. But, as
@swiniker has hinted [4], that code remains inoperable.

In order to clean up our code base, this commit gets rid of these
remainders altogether; particularly it removes function
ilSurveyEvaluationGUI::callPdfGeneration. This changes public API!

It is fairly safe to do so, mainly because the function had already been
renamed in the commits mentioned above, which obviously affected the
API. It is unlikely that plugins have been developed, or adjusted to the
change, which use the old and obsolete functionality under a new name.

Implications: PHP will raise a fatal error when the (now) missing
function is called. (If we kept the function, we would make it throw an
exception anyway to signal that the functionality is gone.)

[0] https://phantomjs.org/
[1] https://github.com/ariya/phantomjs/issues/15344
[2] https://docu.ilias.de/goto_docu_wiki_wpage_5730_1357.html
[3] https://mantis.ilias.de/view.php?id=28709
[4] https://mantis.ilias.de/view.php?id=28709#c68937