0029692: Fix the uncatched division by 0 error
The FormulaQuestion throws an uncatched error (division by 0) in case
"No Restrictions" are set for the "Result Type Selection". This was
fixed for "Fraction" and "Coprime Fraction" in commit
b9b74b660dc7aee0b55dba962e2f32ab8937fa91 and this solution is now
extended to "No Restrictions". This fix was adjusted though, because in
case of a ilMathDivisionByZeroException the users result was set to "0"
which might be a correct answer. Now the division by 0 lets the check
function return false, except in case the formula also threw this
exception.

Mantis-Ticket: 0029692