ilServer: set exit status of process
Use the exit value of the JVM process to signal success or failure to
the caller.

It does not matter if System.exit throws a Throwable as a result of e.g.
super strict security rules; but for the usual case that it does not,
callers can easily react to errors.