HTTP: make ArrayBasedRequestWrapper::retrieve return null on unknown key
Before this commit the class has thrown an exception when an unknown key was used.
This was undocumented in the interface and required people to use has extensively

if ($wrapper->has($key)) {
    $value = $wrapper->retrieve($key, $some_trafo);
}
else {
    $value = $default;
}

Due to the various possibilities of the refinery, this pattern can be completly
internalized like this:

$value = $wrapper->retrieve(
    $key,
     $refinery->byTrying([
        $some_trafo,
        $refinery->always($default)
    ])
);

People could still use the former pattern, though.