T&A: use short form of grade in test results table (Mantis #31209)
Change the “Results and Grades” table in the test results tab so that it
shows the short form of grades instead of the official form.

This harmonizes the displayed marks with the current forms shown by the
statistics table and used in the results export file.

Warning:
changes the form of the data in the ‘finalMark’ field in scoring
objects (instances of class ‘ilTestParticipantScoring’). Although this
is probably a good thing, it it a change of data provided through a
public interface, and code using it (e.g. in plugins) might require
adaptation.

The modified code is also utilized by the SOAP web interface for its
implementation of getTestResults, but that SOAP code does not use the
marks so it should not be negatively affected.

Fixes <https://mantis.ilias.de/view.php?id=31209>.