Auth/PublicArea: Fix auto SSO for SOAP and Apache authentication
With 871004d45b54647f4b683e674585a6baf19d217c the `cmd=force_login`
parameter will always be used if a server request results in the
presentation of the login page. So everytime the login page is
displayed, this parameter is set, even if the `Public Area`
is disabled.
Because of this, the automatic login attempts for `SOAP` and
`Apache` authentication (`ilStartUpGUI::showLoginPage`) do not work
anymore.
This change introduces a further check before
`ilInitialisation::redirectToStartingPage` is called. Without this
change `// fallback, should never happen` in
`ilInitialisation::redirectToStartingPage` does always happen.