fix 35100 by adding a empty variable
ilTestPassOverviewTableGUI line 126 expects the scored variable in the $a_set array for a method call to decide upon an action. By default this scored variable is not set in the array in the getPassOverviewTableData methode. So I initalized it with null. As far as I can tell this should not influence any other classes.

Also Cleanup in the getPassOverviewTableData method