Prevent possible edge case
So !null equals to 1 for PHP, the default for  $isScoredPass is now null, so this could become a problem. However, this check fixes the possible issue.
Furthermore, the relevant call for this code path: IlTestEvaluationGUI line 1473, sets the scoring for the test pass, and therefore it is currently not an issue.