T&A 35100 review of scored answers causes error  (#5169)
* fix 35100 by adding a empty variable

ilTestPassOverviewTableGUI line 126 expects the scored variable in the $a_set array for a method call to decide upon an action. By default this scored variable is not set in the array in the getPassOverviewTableData methode. So I initalized it with null. As far as I can tell this should not influence any other classes.

Also Cleanup in the getPassOverviewTableData method

* Prevent possible edge case

So !null equals to 1 for PHP, the default for  $isScoredPass is now null, so this could become a problem. However, this check fixes the possible issue.
Furthermore, the relevant call for this code path: IlTestEvaluationGUI line 1473, sets the scoring for the test pass, and therefore it is currently not an issue.

* Cleanup or Boy Scouts commit

getPassOverviewTableData takes 3 parameters, not 4

* [fix] Doc block, array syntax and nullable boolean

* [fix] Typisierung entfernt

* [add] Necessary typing