OrgU: Repo Pattern for Positions and Authorities, incl. fixes for 35470,35741,35742,35746
Switched both objects from ActiveRecord to Repo Pattern and edited
all affected parts of OrgU. A change in Services/AccessControl was
necessary due to the new structure, also fixed a small bug in the
handling of the positions' permission data while saving.

A bug in the forms for creating/editing authorities was fixed that
prevented the removal of the first authority and the creation of a
position without authorities.
See Mantis 35470,35741,35742,35746 for bugfix details.