SCORM/Page Editor: Localize word in Long Menu question (Mantis #35064)
When questions are embedded as page content (in terms of the Page
Editor), a bit of client-side JavaScript is utilized that resides in the
Scorm2004 module.

In case of Long Menu questions the handling code lacked localization of
the separator between multiple correct answers in the solution: it was
hardwired to the en-GB string ‘ or ’. This caused English text to appear
in the web UI even though the user language had been set differently. It
happened when the (two or more) correct solutions to a Long Menu
questions were shown.

This commit fixes the issue.

Long Menu questions in tests are not affected (neither by the defect nor
by this commit) because they use a different code path (in PHP code on
server side, not in JavaScript code sent to the client) when answers are
displayed, and that code path already includes localization of the text.

The newly introduced variable name ‘correct_answers_separator’ is a bit
unfortunate: that separator will be surrounded by spaces so it is not a
true separator on its own. When, at some time in future, the text shall
be modified (to e.g. list the first n-1 correct answers separated by
comma+space), both, the code in question_handling.js _and_ in
class.assLongMenu.php, need to be changed in order to remain consistent.

All code utilizing question_handling.js from the Scorm2004 module is
affected and requires a small extension for proper localization: it must
set 'correct_answers_separator' in the JavaScript code of the generated
page. As long as this is not done, an English fallback string will be
used.

Fixes Mantis #35064 <https://mantis.ilias.de/view.php?id=35064>.