T&A LAC: declare types in exception constructors
Add type declarations to the arguments of the constructors of the
LogicalAnswerCompare exceptions. This results in fail-fast behaviour
when instances of them are improperly created. They would probably fail
at a later time anyway.

Some of the declared (non-nullable) types are not yet correct and should
instead be nullable. They have been chosen because part of the existing
internal code (function ‘getQuestionIndex’) uses the non-nullable type.
Future commits are going to fix this issue. For now, the code will fail
at runtime as before, just in another line.

The change aims to increase code readability and help automated tools
which work on the code.