ORGU: fixed various queries in ilOrgUnitPositionAccess
This corrects various broad queries in ilOrgUnitPositionAccess. The queries in
question are supposed to filter groups of users based on permissions and positions
in the Org-Units. They are used in other components to decide which other users
a given user can see.

Before this commit, these queries did return all users when the positions where
not active for a given context or no according operations where found. This caused
https://mantis.ilias.de/view.php?id=36054, because one query was used for a
permission check that was successfull since the position feature was not activated
at that time.

The previous behaviour, returning everyone when position could not be used for
checks, is antithetical to the expected behaviour for permission checks: per default,
user can do nothing (can act on no one), permissions add up to yield more available
actions (or users to act on). The behaviour implied the opposite: If the positions
are not activated it is ok to act on everyone.

In fact, the problematic paths seem to have been triggered rarely: For
filterUserIdsByRbacOrPositionOfCurrentUser (e.g.) there normally would not be a
reason to see any list at all if a user did not have the RBAC-permission or the
positions have not been activated for a given object. Hence, this change is
expected to cause no trouble.

On the other hand, the queries are used to enforce security properties of the system.
The previous behaviour is surprising and may yield to much permissions for a given
scenario (as in the ticket). After the change there might be less permissions then
before, which is either correct or does at least not pose a security risk.