T&A LAC: adjust a type stored in an exception
Stricten a type of a value stored in an exception. (When considering
type juggling, this can be seen as narrowing the type declarations.) The
single place where this exception is used, function ‘validateSubTree’ in
class ‘ilAssLacCompositeValidator’, always passes an integer value to
the exception.

This commit aims to ease understanding of the code when reading. It
changes the API and also negatively affects reusability in future. As
there is no detailed API documentation for the changed exception, this
is still fine (we can just define it to be specific to this use case and
can add further exceptions when the need arises).