T&A LAC: fix question idx type in exception (3/3)
The exception code can refer to a specific question, or be generic and
not give details about the question. In the latter case, in which the
question number is not known (or is not necessary), the code terms it
(and especially its user-centric messages term it) the “current
question”, and the ‘$question_index’ property is null. This null value
is also expected to be passed by the user.

When points for a competence shall be awarded based on an expression,
and that expression refers to a question or an answer in a question, it
is always the current question. The full code path is neither enabled
nor required to handle other questions in this context as that would not
make sense. This means that ‘ilAssLacCompositeValidator’ commonly passes
null as ‘$question_index’.

So the type in the exception must be declared as nullable, which this
commit rectifies.

In case of this commit, the user is function
‘checkAnswerExpressionExist’ in class ‘ilAssLacCompositeValidator’.

Also adjust that function accordingly as we are already there. The
function is private so this should prevent, rather than cause,
additional harm. At least we are going to fail earlier in case of
further defects, which is a good thing.

This commit does not yet change the other use of exception
‘ilAssLacExpressionNotSupportedByQuestion’: this constructor call is
broken and incompatible to the exception code (it would work with string
instead of integer, but this is not good design). A future commit is
necessary to fix that issue.