T&A LAC: add missing language strings
Several language strings used by Test&Assessment LogicalAnswerCompare
exceptions were missing. Add these.

Users will now see meaningful messages when they enter invalid
expressions for competence assignments.

This commit only takes care of languages en and de.

The text for ‘ass_lac_question_not_reachable’ is not a very good
explanation, because the details of the purpose of the using exception
class are not known. This exception, however, is not used at all and so
it does not matter; it will be removed in a future commit.