T&A LAC: split out answer index in exception
Clean up one code path which breaks strong typing and which formerly
used the unfortunate string type and mashed up several things in one
exception property.

Introduce a separate property for the answer index. It defaults to null,
because it is not needed in all cases. Adjust the explanatory strings
accordingly, and add new language strings for en and de which express
the problem and should be shown to users. The language strings contain
shorter texts than the explanation of the exception due to the location
where they are currently used: a short message next to the field with
the erroneous input, in the UI for editing competence assignments.

This extends the API, in a compatible way.