T&A LAC: fix loophole in bracket parsing code
Not only count the number of opening and closing brackets, but also
ensure that they nest properly. Add a new exception for reporting
improper bracket order.

Rewrite the test in ‘ilAssLacConditionParser->checkBrackets’ to fix the
issue. The new version should be a bit more efficient than the trivial
addition of a nesting check: only in the case of a missing opening
bracket count all remaining brackets. This means we only loop over the
string once. (Anyway, this is neglegible for all real-world scenarios.)

New language strings are only added for en and de.