T&A LAC: get types right on composite building
Straighten out the code affected by types where a nodes tree from the
parser is converted to a proper composite tree object structure: make it
less lazy and more explicit, thereby strengthening consistency in types
usage.

Add more API documentation for the affected method, especially to
clarify its input expectations. Passing it unexpected input may cause it
to crash.

In one of the erroneous cases, namely when the passed node array
structure does not represent a node group, the composite building
function now explicitly throws an exception. This changes the function
API, but is still considered safe to introduce, as the function would
fail (and even worse) anyway.

The composite builder is used by the LAC condition parser. So this
commit repairs the type of the root of the object tree returned by that
parser.

Fixes the follow-up problems detected after getting rid of the cause of
<https://mantis.ilias.de/view.php?id=35685>.