Miscellaneous small assessment improvements (#5395)
* T&A API doc: document meaning of a skills-related function

* T&A: simplify a small piece of skills-related code

Use a function already provided by class ‘ilObjTest’. That function is
already used earlier in the code.

* T&A API doc: explain purpose of some skills-related classes

Document these in order to help people understand the code.

Three of the classes are the (only) entry points to (= users of)
LogicalAnswerCompare code (the skill evaluation GUI class is not, but
documenting it as well is handy to understand its non-GUI companion).

* T&A: use proper boolean operators on booleans

Change odd uses of operators: combine boolean values with boolean
operators instead of bitwise operators. The former code has been working
due to type juggling, but is not clean and we can avoid it.

* T&A: simplify code by relying on class autoloading

Remove a few occurrences of explicit class loading (explicitly
require-ing of another script). They are not needed because the
autoloader will include the respective script automatically.

This commit makes the changed function in each affected class consistent
with its equivalents in other question classes: they already utilize
autoloading there. It also simplifies the code, enhancing readability by
a bit.